agents/nj_voter_chat_adk/requirements.txt
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
google-cloud-bigquery>=3.25.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-api-core>=2.19.0
google-adk>=0.1.0
vertexai>=1.38.0
python-dotenv>=1.0.1
streamlit>=1.36.0
pytest>=8.2.0
requests>=2.31.0
